+++ FluentFTP.CSharpExamples/GetChecksum.cs
@@ -1,0 +1,46 @@
+﻿using System;
+using System.Net;
+using System.Threading;
+using System.Threading.Tasks;
+using FluentFTP;
+
+namespace Examples {
+	public static class GetChecksumExample {
+
+		public static void GetChecksum() {
+			
+			using (var conn = new FtpClient("127.0.0.1", "ftptest", "ftptest")) {
+				conn.Connect();
+
+				// Get a hash checksum for the file
+				FtpHash hash = conn.GetChecksum("/path/to/remote/file");
+
+				// Make sure it returned a valid hash object
+				if (hash.IsValid) {
+					if (hash.Verify("/some/local/file")) {
+						Console.WriteLine("The checksum's match!");
+					}
+				}
+			}
+		}
+
+
+		public static async Task GetChecksumAsync() {
+			var token = new CancellationToken();
+			using (var conn = new FtpClient("127.0.0.1", "ftptest", "ftptest")) {
+				await conn.ConnectAsync(token);
+
+				// Get a hash checksum for the file
+				FtpHash hash = await conn.GetChecksumAsync("/path/to/remote/file", token);
+
+				// Make sure it returned a valid hash object
+				if (hash.IsValid) {
+					if (hash.Verify("/some/local/file")) {
+						Console.WriteLine("The checksum's match!");
+					}
+				}
+			}
+		}
+
+	}
+}